remove circular dep between model/effects and pokemon-entity
