+++ app/core/pokemon-entity.ts
@@ -7,14 +7,13 @@ import {
   MAX_SPEED,
   ON_ATTACK_MANA
 } from "../config"
+import { SynergyEffects } from "../config/game/synergies"
 import Count from "../models/colyseus-models/count"
 import Player from "../models/colyseus-models/player"
 import { type Pokemon, PokemonClasses } from "../models/colyseus-models/pokemon"
 import Status from "../models/colyseus-models/status"
-import { SynergyEffects } from "../models/effects"
 import PokemonFactory from "../models/pokemon-factory"
 import { getPokemonData } from "../models/precomputed/precomputed-pokemon-data"
-import { getSellPrice } from "../models/shop"
 import {
   Emotion,
   type IPokemon,
@@ -49,7 +48,7 @@ import { isOnBench } from "../utils/board"
 import { distanceC, distanceM } from "../utils/distance"
 import { isPlainFunction } from "../utils/function"
 import { clamp, min, roundToNDigits } from "../utils/number"
-import { chance, pickNRandomIn, pickRandomIn } from "../utils/random"
+import { chance, pickNRandomIn } from "../utils/random"
 import { schemaValues } from "../utils/schemas"
 import AttackingState from "./attacking-state"
 import type { Board } from "./board"
@@ -79,7 +78,6 @@ import MovingState from "./moving-state"
 import type PokemonState from "./pokemon-state"
 import type Simulation from "./simulation"
 import { DelayedCommand, type SimulationCommand } from "./simulation-command"
-import { getUnitScore } from "./unit-score";
 
 export class PokemonEntity extends Schema implements IPokemonEntity {
   @type("boolean") shiny: boolean
